tournament-server/requirements.txt
alembic==1.10.3
autopep8==2.0.2
bcrypt==4.0.1
blinker==1.6.2
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
cryptography==40.0.2
dnspython==2.3.0
email-validator==2.0.0.post1
Flask==2.2.3
Flask-Bcrypt==1.0.1
Flask-Cors==3.0.10
Flask-DebugToolbar==0.13.1
Flask-Migrate==4.0.4
Flask-SQLAlchemy==3.0.3
greenlet==2.0.2
idna==3.4
imageio==2.27.0
imagekitio==3.0.1
itsdangerous==2.1.2
Jinja2==3.1.2
jwt==1.3.1
Mako==1.2.4
MarkupSafe==2.1.2
numpy==1.24.2
Pillow==9.5.0
pycodestyle==2.10.0
pycparser==2.21
PyJWT==1.7.1
python-dotenv==1.0.0
requests==2.28.2
requests-toolbelt==0.9.1
six==1.16.0
SQLAlchemy==2.0.9
tomli==2.0.1
typing_extensions==4.5.0
urllib3==1.26.15
Werkzeug==2.2.3
